apps/workers/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CrewAI and AI/ML
crewai==0.1.0
openai==1.3.0
anthropic==0.7.0
langchain==0.0.350
langchain-openai==0.0.2
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4
scikit-learn==1.3.2
networkx==3.2.1

# Visualization and reporting
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
reportlab==4.0.7

# Database and storage
asyncpg==0.29.0
redis==5.0.1
clickhouse-driver==0.2.6
boto3==1.34.0
sqlalchemy==2.0.23
alembic==1.13.0

# Messaging and events
nats-py==2.6.0
celery==5.3.4

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
httpx==0.25.2

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
typer==0.9.0
rich==13.7.0
